seance 11

+++ my-blog/routes/web.php
@@ -20,18 +20,21 @@ Route::get('/', function () {
 
 
 
-Route::get('blog', [BlogPostController::class, 'index'])->name('blog.index');
-Route::get('blog/{blogPost}', [BlogPostController::class, 'show'])->name('blog.show');
-Route::get('blog-create', [BlogPostController::class, 'create'])->name('blog.create');
-Route::post('blog-create', [BlogPostController::class, 'store'])->name('blog.store');
-Route::get('blog-edit/{blogPost}', [BlogPostController::class, 'edit'])->name('blog.edit');
-Route::put('blog-edit/{blogPost}', [BlogPostController::class, 'update'])->name('blog.update');
-Route::delete('blog-edit/{blogPost}', [BlogPostController::class, 'destroy'])->name('blog.delete');
+Route::get('blog', [BlogPostController::class, 'index'])->name('blog.index')->middleware('auth');
+Route::get('blog/{blogPost}', [BlogPostController::class, 'show'])->name('blog.show')->middleware('auth');
+Route::get('blog-create', [BlogPostController::class, 'create'])->name('blog.create')->middleware('auth');
+Route::post('blog-create', [BlogPostController::class, 'store'])->name('blog.store')->middleware('auth');
+Route::get('blog-edit/{blogPost}', [BlogPostController::class, 'edit'])->name('blog.edit')->middleware('auth');
+Route::put('blog-edit/{blogPost}', [BlogPostController::class, 'update'])->name('blog.update')->middleware('auth');
+Route::delete('blog-edit/{blogPost}', [BlogPostController::class, 'destroy'])->name('blog.delete')->middleware('auth');
 
 Route::get('query', [BlogPostController::class, 'query']);
 
-Route::get('blog-page', [BlogPostController::class, 'page']);
+Route::get('blog-page', [BlogPostController::class, 'page'])->middleware('auth');
 
 Route::get('registration', [CustomAuthController::class, 'create'])->name('user.registration');
 Route::post('registration', [CustomAuthController::class, 'store']);
-Route::get('user-list', [CustomAuthController::class, 'userList'])->name('user.list');
+Route::get('user-list', [CustomAuthController::class, 'userList'])->name('user.list')->middleware('auth');
+Route::get('login', [CustomAuthController::class, 'index'])->name('login');
+Route::post('login', [CustomAuthController::class, 'authentication']);
+Route::get('logout', [CustomAuthController::class, 'logout'])->name('logout')->middleware('auth');
